requirements.txt
pyspark>=3.3.0
delta-spark>=2.2.0
boto3>=1.26.0
requests>=2.28.0
python-dotenv>=1.0.0
